Reorganize num related codes
