+++ nmf/__init__.py
@@ -1,15 +1,15 @@
-from .nmf import run_nmf
+# from .nmf import run_nmf
 
-from ._nmf_batch import NMFBatch
+from ._nmf_batch_mu import NMFBatchMU
 from ._nmf_batch_hals import NMFBatchHALS
-from ._nmf_online import NMFOnline
+from ._nmf_online_mu import NMFOnlineMU
 from ._nmf_online_hals import NMFOnlineHALS
 
-from ._inmf_batch_mu import INMFBatchMU
-from ._inmf_batch_hals import INMFBatchHALS
-# from ._inmf_batch_nnls_bpp import INMFBatchNnlsBpp
-from ._inmf_online_mu import INMFOnlineMU
-from ._inmf_online_hals import INMFOnlineHALS
+# from ._inmf_batch_mu import INMFBatchMU
+# from ._inmf_batch_hals import INMFBatchHALS
+# # from ._inmf_batch_nnls_bpp import INMFBatchNnlsBpp
+# from ._inmf_online_mu import INMFOnlineMU
+# from ._inmf_online_hals import INMFOnlineHALS
 
 try:
     from importlib.metadata import version, PackageNotFoundError
