+++ nmf/_inmf_online_mu.py
@@ -4,40 +4,6 @@ from ._inmf_online_base import INMFOnlineBase
 from typing import List, Union
 
 class INMFOnlineMU(INMFOnlineBase):
-    def __init__(
-        self,
-        n_components: int,
-        lam: float = 5.,
-        init: str = 'random',
-        tol: float = 1e-4,
-        random_state: int = 0,
-        fp_precision: Union[str, torch.dtype] = 'float',
-        device_type: str = 'cpu',
-        max_pass: int = 20,
-        chunk_size: int = 5000,
-        max_iter: int = 200,
-        h_tol: float = 0.01,
-        v_tol: float = 0.1,
-        w_tol: float = 0.01,
-    ):
-        super().__init__(
-            n_components=n_components,
-            lam=lam,
-            init=init,
-            tol=tol,
-            random_state=random_state,
-            fp_precision=fp_precision,
-            device_type=device_type,
-        )
-
-        self._max_pass = max_pass
-        self._chunk_size = chunk_size
-        self._max_iter = max_iter
-        self._h_tol = h_tol
-        self._v_tol = v_tol
-        self._w_tol = w_tol
-
-
     def _update_matrix(self, mat, numer, denom):
         rates = numer / denom
         rates[denom < self._epsilon] = 0.0
@@ -77,10 +43,10 @@ class INMFOnlineMU(INMFOnlineBase):
                 xWVT = x @ WV.T
 
                 h_factor_numer = xWVT
-                for j in range(self._max_iter):
+                for j in range(self._chunk_max_iter):
                     h_factor_denom = h @ (WVWVT + self._lambda * VVT) if self._lambda > 0.0 else h @ WVWVT
                     cur_max = self._update_matrix(h, h_factor_numer, h_factor_denom)
-                    if j + 1 < self._max_iter and cur_max / h.mean() < self._h_tol:
+                    if j + 1 < self._chunk_max_iter and cur_max / h.mean() < self._h_tol:
                         break
                 # print(f"Batch {k} Block {i} update H iterates {j+1} iterations.")
                 self.H[k][idx, :] = h
@@ -93,11 +59,11 @@ class INMFOnlineMU(INMFOnlineBase):
 
                 # Update V
                 V_factor_numer = B
-                for j in range(self._max_iter):
+                for j in range(self._chunk_max_iter):
                     V_factor_denom = A @ (WV + self._lambda * self.V[k])
                     cur_max = self._update_matrix(self.V[k], V_factor_numer, V_factor_denom)
                     WV = self.W + self.V[k]
-                    if j + 1 < self._max_iter and cur_max / self.V[k].mean() < self._v_tol:
+                    if j + 1 < self._chunk_max_iter and cur_max / self.V[k].mean() < self._v_tol:
                         break
                 # print(f"Batch {k} Block {i} update V iterates {j+1} iterations.")
 
@@ -108,10 +74,10 @@ class INMFOnlineMU(INMFOnlineBase):
 
                 # Update W
                 W_factor_numer = D
-                for j in range(self._max_iter):
+                for j in range(self._chunk_max_iter):
                     W_factor_denom = C @ self.W + E_new
                     cur_max = self._update_matrix(self.W, W_factor_numer, W_factor_denom)
-                    if j + 1 < self._max_iter and cur_max / self.W.mean() < self._w_tol:
+                    if j + 1 < self._chunk_max_iter and cur_max / self.W.mean() < self._w_tol:
                         break
                 # print(f"Batch {k} Block {i} update W iterates {j+1} iterations.")
                 i += self._chunk_size
@@ -145,10 +111,10 @@ class INMFOnlineMU(INMFOnlineBase):
                 xWVT = x @ WV.T
 
                 h_factor_numer = xWVT
-                for j in range(self._max_iter):
+                for j in range(self._chunk_max_iter):
                     h_factor_denom = h @ (WVWVT + self._lambda * VVT) if self._lambda > 0.0 else h @ WVWVT
                     cur_max = self._update_matrix(h, h_factor_numer, h_factor_denom)
-                    if j + 1 < self._max_iter and cur_max / h.mean() < self._h_tol:
+                    if j + 1 < self._chunk_max_iter and cur_max / h.mean() < self._h_tol:
                         break
                 # print(f"Batch {k} Block {i} update H iterates {j+1} iterations.")
                 self.H[k][idx, :] = h
@@ -161,11 +127,11 @@ class INMFOnlineMU(INMFOnlineBase):
 
                 # Update V
                 V_factor_numer = B
-                for j in range(self._max_iter):
+                for j in range(self._chunk_max_iter):
                     V_factor_denom = A @ (WV + self._lambda * self.V[k])
                     cur_max = self._update_matrix(self.V[k], V_factor_numer, V_factor_denom)
                     WV = self.W + self.V[k]
-                    if j + 1 < self._max_iter and cur_max / self.V[k].mean() < self._v_tol:
+                    if j + 1 < self._chunk_max_iter and cur_max / self.V[k].mean() < self._v_tol:
                         break
                 # print(f"Batch {k} Block {i} update V iterates {j+1} iterations.")
                 i += self._chunk_size
@@ -173,7 +139,7 @@ class INMFOnlineMU(INMFOnlineBase):
 
     def _update_H(self):
         """ Fix W and V, update H """
-        sum_h_err = torch.tensor(0.0, dtype = torch.double) # make sure sum_h_err is double to avoid summation errors
+        sum_h_err = torch.tensor(0.0, dtype=torch.double, device=self._device_type) # make sure sum_h_err is double to avoid summation errors
         for k in range(self._n_batches):
             WV = self.W + self.V[k]
             WVWVT = WV @ WV.T
@@ -187,10 +153,10 @@ class INMFOnlineMU(INMFOnlineBase):
                 # Update H
                 xWVT = x @ WV.T
                 h_factor_numer = xWVT
-                for j in range(self._max_iter):
+                for j in range(self._chunk_max_iter):
                     h_factor_denom = h @ (WVWVT + self._lambda * VVT) if self._lambda > 0.0 else h @ WVWVT
                     cur_max = self._update_matrix(h, h_factor_numer, h_factor_denom)
-                    if j + 1 < self._max_iter and cur_max / h.mean() < self._h_tol:
+                    if j + 1 < self._chunk_max_iter and cur_max / h.mean() < self._h_tol:
                             break
                 # print(f"Batch {k} Block {i} update H iterates {j+1} iterations.")
